// backoffice/src/TypesFromApi/models/Role.ts
/* tslint:disable */
/* eslint-disable */
/**
 * pass Culture backoffice API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime'
import {
  Permission,
  PermissionFromJSON,
  PermissionFromJSONTyped,
  PermissionToJSON,
} from './index'

/**
 *
 * @export
 * @interface Role
 */
export interface Role {
  /**
   *
   * @type {number}
   * @memberof Role
   */
  id: number
  /**
   *
   * @type {string}
   * @memberof Role
   */
  name: string
  /**
   *
   * @type {Array<Permission>}
   * @memberof Role
   */
  permissions: Array<Permission>
}

export function RoleFromJSON(json: any): Role {
  return RoleFromJSONTyped(json, false)
}

export function RoleFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): Role {
  if (json === undefined || json === null) {
    return json
  }
  return {
    id: json['id'],
    name: json['name'],
    permissions: (json['permissions'] as Array<any>).map(PermissionFromJSON),
  }
}

export function RoleToJSON(value?: Role | null): any {
  if (value === undefined) {
    return undefined
  }
  if (value === null) {
    return null
  }
  return {
    id: value.id,
    name: value.name,
    permissions: (value.permissions as Array<any>).map(PermissionToJSON),
  }
}